Scripts for management of big number of files
This two scripts can be used for sorting files in multiple  directories
1) list_XISF_files.py - creates a list of files (with a certain extension or/and naming pattern) in a given directory
2) copy-move_listed_files.py - copies or moves files matching the list created in a previous step, from 'Directory 1' to 'Directory 2'.
Both scripts should be adjusted before use  (i.e. the user is supposed to provide a correct path where the scripts are to be executed, as well as the file extension or/and the name fragment for the files to be listed, and the names of source_dir and dest_dir). Those scripts employ os and shutil modules